chore: remove conditional frontend routing (#6190)
* chore: remove react migration vars

* chore: remove more conditional code in angular, backend types

* chore: remove angular from express routing rules!

* chore: add back public route

* chore: remove angular public route

* chore: add react router

* chore: delete home folder, rename frontend to frontend-old and frontend-new to frontend

* chore: remove src_e2e ci test, rename tests to better reflect backend-frontend split

* test: add --openssl-legacy-provider to the ci build step

* chore: update node options for build step

* chore: revert ci to node 14
