+++ package.json
@@ -25,30 +25,22 @@
   },
   "scripts": {
     "postinstall": "npm run postinstall:frontend && npm run postinstall:shared",
+    "test": "npm run test:backend && npm run test:frontend",
     "test:backend": "env-cmd -f tests/.test-env jest --coverage --maxWorkers=4",
     "test:backend:watch": "env-cmd -f tests/.test-env jest --watch",
     "test:frontend": "npm --prefix frontend test",
     "test:e2e-v2": "npm run build && npx playwright test",
     "test:e2e-v2:server": "env-cmd -f ./__tests__/e2e/setup/.test-env npm run test-e2e-server",
+    "test-e2e-server": "concurrently --success last --kill-others \"mockpass\" \"maildev\" \"node dist/backend/src/app/server.js\" \"node ./tests/mock-webhook-server.js\"",
     "build": "npm run clean && concurrently \"npm:build:*\"",
-    "test:angularjs": "jest --config=tests/unit/frontend/jest.config.js",
-    "dev:angularjs": "webpack --config webpack.dev.js",
-    "build:old-frontend": "webpack --config webpack.prod.js",
     "build:frontend": "npm run --prefix frontend build",
     "build:backend": "tsc -p tsconfig.build.json && npm run copyfiles:backend",
     "copyfiles:backend": "copyfiles -e src/public/**/*.html -u 1 src/**/*.html dist/backend/src",
     "clean": "rimraf dist/",
     "start": "node -r dotenv/config src/app/server.js",
     "dev": "concurrently -k -p \"[{name}]\" -n \"api,client\" -c  \"yellow.bold,green.bold\" \"docker-compose up\" \"npm run dev:frontend\"",
-    "dev:backend": "npm run dev:angularjs & tsnd --poll --respawn --transpile-only --inspect=0.0.0.0 --exit-child -r dotenv/config -- src/app/server.ts",
+    "dev:backend": "tsnd --poll --respawn --transpile-only --inspect=0.0.0.0 --exit-child -r dotenv/config -- src/app/server.ts",
     "dev:frontend": "npm run --prefix frontend start",
-    "test": "npm run test:backend && npm run test:frontend",
-    "download-binary": "node tests/end-to-end/helpers/get-mongo-binary.js",
-    "test-e2e": "npm run test-e2e-build && npm run test-e2e-ci",
-    "test-e2e-build": "npm run build:backend && npm run dev:angularjs",
-    "test-e2e-ci": "env-cmd -f tests/.test-env --use-shell \"npm run download-binary && npm run testcafe-command\"",
-    "testcafe-command": "testcafe --skip-js-errors -c 2 chrome:headless ./tests/end-to-end --app \"npm run test-e2e-server\" --app-init-delay 10000",
-    "test-e2e-server": "concurrently --success last --kill-others \"mockpass\" \"maildev\" \"node dist/backend/src/app/server.js\" \"node ./tests/mock-webhook-server.js\"",
     "lint-code": "eslint src/ --quiet --fix",
     "lint-style": "stylelint '*/**/*.css' --quiet --fix",
     "lint-html": "htmlhint && prettier --write './src/public/**/*.html' --ignore-path './dist/**' --loglevel silent",
