+++ shared/types/core.ts
@@ -32,12 +32,6 @@ export type ClientEnvVars = {
   adminBannerContentReact: string
 
   spcpCookieDomain: string // Cookie domain used for removing spcp cookies
-  respondentRolloutEmail: number
-  respondentRolloutStorage: number
-  adminRollout: number
-  angularPhaseOutDate: string | null
-  removeAdminInfoboxThreshold: number
-  removeRespondentsInfoboxThreshold: number
   stripePublishableKey: string
   maxPaymentAmountCents: number
   minPaymentAmountCents: number
