+++ shared/types/index.ts
@@ -4,7 +4,6 @@ export * from './agency'
 export * from './analytics'
 export * from './billing'
 export * from './core'
-export * from './env'
 export * from './generic'
 export * from './response'
 export * from './submission'
