+++ src/app/loaders/express/index.ts
@@ -12,10 +12,9 @@ import { ExamplesRouter } from '../../modules/examples/examples.routes'
 import { AdminFormsRouter } from '../../modules/form/admin-form/admin-form.routes'
 import { PublicFormRouter } from '../../modules/form/public-form/public-form.routes'
 import { FrontendRouter } from '../../modules/frontend/frontend.routes'
-import * as HomeController from '../../modules/home/home.controller'
+import { FrontendRouter as OldFrontendRouter } from '../../modules/frontend-old/frontend.routes'
 import { MYINFO_ROUTER_PREFIX } from '../../modules/myinfo/myinfo.constants'
 import { MyInfoRouter } from '../../modules/myinfo/myinfo.routes'
-import { ReactMigrationRouter } from '../../modules/react-migration/react-migration.routes'
 import { SgidRouter } from '../../modules/sgid/sgid.routes'
 import { SubmissionRouter } from '../../modules/submission/submission.routes'
 import { VfnRouter } from '../../modules/verification/verification.routes'
@@ -114,7 +113,7 @@ const loadExpressApp = async (connection: Connection) => {
   app.use(IntranetMiddleware.logIntranetUsage)
 
   // Deprecated routes
-  app.use('/frontend', FrontendRouter)
+  app.use('/frontend', OldFrontendRouter)
   app.use('/auth', AuthRouter)
   app.use('/transaction', VfnRouter)
   app.use('/examples', ExamplesRouter)
@@ -137,7 +136,6 @@ const loadExpressApp = async (connection: Connection) => {
   // serve static assets. `dist/frontend` contains the root files as well as a `/static` folder
   // express.static calls next() if the file is not found
   app.use(express.static(path.resolve('dist/frontend'), { index: false }))
-  app.use('/public', express.static(path.resolve('dist/angularjs')))
 
   // If requests for known static asset patterns were not served by
   // the static handlers above, middleware should try to fetch from s3 static bucket or else return 404s
@@ -147,8 +145,7 @@ const loadExpressApp = async (connection: Connection) => {
   // not served statically above will also return 404
   app.get(/^\/[^/]+\.[a-z]+$/, catchNonExistentStaticRoutesMiddleware)
 
-  app.get('/old/', HomeController.home)
-  app.use('/', ReactMigrationRouter)
+  app.use('/', FrontendRouter)
 
   app.use(sentryMiddlewares())
   app.use(errorHandlerMiddlewares())
