+++ src/app/modules/feedback/feedback.controller.ts
@@ -2,8 +2,6 @@ import { celebrate, Joi, Segments } from 'celebrate'
 import { StatusCodes } from 'http-status-codes'
 import { ErrorDto, PrivateFormErrorDto } from 'shared/types'
 
-import config from '../../config/config'
-import { statsdClient } from '../../config/datadog-statsd-client'
 import { createLoggerWithLabel } from '../../config/logger'
 import { createReqMeta } from '../../utils/request'
 import { ControllerHandler } from '../core/core.types'
@@ -56,16 +54,8 @@ const submitFormFeedback: ControllerHandler<
     .andThen(() => FeedbackService.hasNoPreviousFeedback(submissionId))
     .andThen(() => FormService.retrieveFullFormById(formId))
     .andThen((form) => FormService.isFormPublic(form).map(() => form))
-    .andThen((form) => {
-      statsdClient.distribution('formsg.feedback.rating', rating, 1, {
-        rating: `${rating}`,
-        ui:
-          req.cookies?.[config.reactMigration.respondentCookieName] === 'react'
-            ? 'react'
-            : 'angular',
-      })
-
-      return PublicFormService.insertFormFeedback({
+    .andThen((form) =>
+      PublicFormService.insertFormFeedback({
         formId: form._id,
         submissionId: submissionId,
         rating,
@@ -74,8 +64,8 @@ const submitFormFeedback: ControllerHandler<
         res
           .status(StatusCodes.OK)
           .json({ message: 'Successfully submitted feedback' }),
-      )
-    })
+      ),
+    )
     .mapErr((error) => {
       const { errorMessage, statusCode } = mapRouteError(error)
       logger.error({
