+++ src/public/main.css
@@ -23,7 +23,6 @@
 @import './modules/core/css/banner.css';
 @import './modules/core/css/how-it-works.css';
 @import './modules/forms/admin/css/pop-up-modal.css';
-@import './modules/forms/admin/css/react-switch-banner.css';
 @import './modules/forms/admin/css/edit-form.css';
 @import './modules/forms/admin/css/export-button.css';
 @import './modules/forms/admin/css/form-title-input.css';
