+++ src/public/main.js
@@ -193,7 +193,6 @@ require('./modules/forms/admin/components/edit-logic.client.component.js')
 require('./modules/forms/admin/components/export-button.client.component.js')
 require('./modules/forms/admin/components/form-title-input.client.component.js')
 require('./modules/forms/admin/components/form-emails-input.client.component.js')
-require('./modules/forms/admin/components/react-switch-banner.client.component.js')
 
 // response components
 require('./modules/forms/admin/components/response-components/response.client.component.js')
@@ -366,10 +365,6 @@ app.run([
     $templateCache.put(
       'modules/forms/admin/componentViews/form-emails-input.client.view.html',
       require('./modules/forms/admin/componentViews/form-emails-input.client.view.html'),
-    )
-    $templateCache.put(
-      'modules/forms/admin/componentViews/react-switch-banner.client.view.html',
-      require('./modules/forms/admin/componentViews/react-switch-banner.client.view.html'),
     )
 
     // Response component views
