--- src/public/modules/forms/admin/componentViews/react-switch-banner.client.view.html
@@ -1,12 +1,0 @@
-<!-- Switch from AngularJS to React message -->
-<div id="react-switch-banner" class="react-switch-banner-container">
-  <img class="info-icon" src="/public/modules/core/img/info-icon.svg" />
-  <span class="is-text">A new FormSG is now available.&nbsp</span>
-  <div
-    ng-click="vm.adminChooseEnvironment(vm.ui)"
-    class="react-switch-banner-container--link"
-    id="react-switch-banner-link"
-  >
-    Try it here!
-  </div>
-</div>
