+++ src/public/modules/forms/admin/controllers/list-forms.client.controller.js
@@ -36,8 +36,6 @@ function ListFormsController(
 
   vm.bannerContent = $window.siteBannerContent || $window.adminBannerContent
 
-  // Admin React rollout percentage
-  vm.reactAdminRollout = $window.reactMigrationAdminRollout
   // Hidden buttons on forms that appear after clicking more
   vm.showFormBtns = []
   // Duplicated form outline on newly dup forms
