+++ src/public/modules/forms/base/controllers/submit-form.client.controller.js
@@ -31,18 +31,6 @@ function SubmitFormController(
   const vm = this
   const form = FormData.form
 
-  // React migration checker - ONLY for plain form URLs (no suffixes like /template or /preview)
-  if (/^\/[0-9a-fA-F]{24}\/?$/.test($location.path())) {
-    const respondentCookie = $cookies.get(
-      $window.reactMigrationRespondentCookieName,
-    )
-
-    if (respondentCookie === 'react') {
-      $window.location.assign(`/${form._id}`)
-      return
-    }
-  }
-
   // The form attribute of the FormData object contains the form fields, logic etc
   vm.myform = form
