+++ src/public/services/PublicFormService.ts
@@ -91,50 +91,3 @@ export const getPublicFormView = async (
     .get<PublicFormViewDto>(`${PUBLIC_FORMS_ENDPOINT}/${formId}`)
     .then(({ data }) => data)
 }
-
-// TODO #4279: Remove after React rollout is complete
-/**
- * Gets public view of the react to angular bug report feedback form
- * @returns Public view of form, with additional identify information
- */
-export const getReactToAngularFeedbackFormView =
-  async (): Promise<PublicFormViewDto> => {
-    return axios
-      .get<PublicFormViewDto>(`${PUBLIC_FORMS_ENDPOINT}/switchenvfeedback`)
-      .then(({ data }) => data)
-  }
-
-/**
- * Submits the bug report feedback form for React to Angular switch
- * @param content content of submission
- * @param attachments any attachments included in submission
- * @param captchaResponse string if captcha to be included, defaults to null
- *
- * @returns SubmissionResponseDto if successful, else SubmissionErrorDto on error
- */
-export const submitSwitchEnvFeedbackForm = async ({
-  content,
-  attachments,
-  captchaResponse = null,
-}: {
-  content: EmailModeSubmissionContentDto
-  attachments?: Record<string, File>
-  captchaResponse?: string | null
-}): Promise<SubmissionResponseDto> => {
-  const formData = createEmailSubmissionFormData({
-    content,
-    attachments,
-  })
-
-  return axios
-    .post<SubmissionResponseDto>(
-      `${PUBLIC_FORMS_ENDPOINT}/submissions/email/switchenvfeedback`,
-      formData,
-      {
-        params: {
-          captchaResponse: String(captchaResponse),
-        },
-      },
-    )
-    .then(({ data }) => data)
-}
